+++ Modules/Microsoft365DSC/DscResources/MSFT_AADAgreement/MSFT_AADAgreement.psm1
@@ -259,6 +259,7 @@ class AADAgreement : M365DSCResourceBase
                     }
                     fileName = $this.FileName
                     language = $this.Language
+                    isDefault = $true
                 }
             }